+++ modules/sagemaker/sagemaker-templates/templates/finetune_llm_evaluation/pipeline_constructs/build_pipeline_construct.py
@@ -228,8 +228,7 @@ class BuildPipelineConstruct(Construct):
                     "sagemaker:ListTags",
                 ],
                 resources=[
-                    f"arn:{Aws.PARTITION}:sagemaker:{Aws.REGION}:{Aws.ACCOUNT_ID}:pipeline/"
-                    f"{sagemaker_pipeline_name}",
+                    f"arn:{Aws.PARTITION}:sagemaker:{Aws.REGION}:{Aws.ACCOUNT_ID}:pipeline/{sagemaker_pipeline_name}",
                     f"arn:{Aws.PARTITION}:sagemaker:{Aws.REGION}:{Aws.ACCOUNT_ID}:pipeline/"
                     f"{sagemaker_pipeline_name}/execution/*",
                 ],