+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### **Added**
 
 ### **Changed**
+- updated GitHub repo creation in `sagemaker-templates` module to support GitHub organizations
 
 ## v2.1.0